+++ src/types/dataTypes/drupal/field_type.d.ts
@@ -64,3 +64,21 @@ export interface FieldAdministration {
   drupal_internal__tid: number
   name: string
 }
+
+/**
+ * Types for breadcrumb data
+ */
+export interface Breadcrumb {
+  url: {
+    path: string
+    routed: boolean
+  }
+  text: string
+}
+
+export interface BreadcrumbItem {
+  uri: string
+  title: string
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  options: any[]
+}
